red_knot: first attempt at string annotation
This separates the inference of string annotations in the deferred
region. But, this creates complications in annotations that are only
partially stringified e.g., `tuple[int, "Foo"]` where "Foo" is a forward
reference.

This commit exists so as to create a checkpoint in case some of the
ideas explored here are useful.