Infer string annotations in the same definition query
This is the second attempt for string annotation which infers the string
annotation types in the same definition query. This has the added
advantage of avoiding to go through two salsa queries. It does this by
maintaining a state on the builder and utilizes that to make certain
decisions throughout the inference process.